removed unnecessary (long) casts in anf.c

+++ anf.c
@@ -22,8 +22,8 @@ int anf(int y, int *s, int *a, int *rho, unsigned int* index)
     long a_i = (long) *a;
 
     // STEP 1): update rho
-    AC0 = (long) lambda * (long) rho[0]; // Q15 * Q15 = Q30
-    AC1 = (long) (32767 - lambda) * (long) rho[1]; // Q15 * Q15 = Q30
+    AC0 = (long) lambda * rho[0]; // Q15 * Q15 = Q30
+    AC1 = (long) (32767 - lambda) * rho[1]; // Q15 * Q15 = Q30
     AC0 += AC1; // Q30 + Q30 = Q30
     AC0 += 16384; // Round the part we'll truncate by adding 2^14
     rho[0] = (int) (AC0 >> 15); // Q30 -> Q15
@@ -32,15 +32,15 @@ int anf(int y, int *s, int *a, int *rho, unsigned int* index)
     AC1 = (((long) rho[0]) >> 1) * a_i; // Q14 * Q14 = Q28
     AC1 += 32768; // Round the part we'll truncate by adding 2^15
     AC1 >>= 16; // Q28 -> Q12
-    AC1 = AC1 * (long) s[k_minus_1]; // Q12 * Q12 -> Q24
+    AC1 = AC1 * s[k_minus_1]; // Q12 * Q12 -> Q24
 
     AC0 = ((long) y) << 9; // Q15 -> Q24
     AC0 += AC1; // Q24
 
-    AC1 = ((long) rho[0]) * ((long) rho[0]); // Q15 * Q15 = Q30
+    AC1 = (long) rho[0] * rho[0]; // Q15 * Q15 = Q30
     AC1 += 131072; // Round the part we'll truncate by adding 2^17
     AC1 >>= 18; // Q30 -> Q12
-    AC1 = AC1 * (long) s[k_minus_2];  //Q12 * Q12 = Q24
+    AC1 = AC1 * s[k_minus_2];  //Q12 * Q12 = Q24
 
     AC0 -= AC1;
     AC0 += 2048; // Round the part we'll truncate by adding 2^11
@@ -48,17 +48,17 @@ int anf(int y, int *s, int *a, int *rho, unsigned int* index)
     s[k] = (int) (AC0 >> 12); // Q24 -> Q12
 
     // STEP 3): update e
-    AC0 = a_i * (long) s[k_minus_1]; // Q14 * Q12 = Q26
+    AC0 = a_i * s[k_minus_1]; // Q14 * Q12 = Q26
     AC0 = (((long) s[k]) << 14) + (((long) s[k_minus_2]) << 14) - AC0; // Q26
     AC0 += 1024; // Round the part we'll truncate by adding 2^10
     e = (int) (AC0 >> 11); // Q26 -> Q15
 
     // STEP 4): update a
-    AC0 = (long) (2 << 13) * (long) mu; // Q13 * Q15 = Q28
+    AC0 = (long) (2 << 13) * mu; // Q13 * Q15 = Q28
     AC0 += 8192; // Round the part we'll truncate by adding 2^13
     AC0 = AC0 >> 14; //Q28->Q14
 
-    AC1 = (long) e * (long) s[k_minus_1]; // Q15 * Q12 = Q27
+    AC1 = (long) e * s[k_minus_1]; // Q15 * Q12 = Q27
     AC1 += 4096; // Round the part we'll truncate by adding 2^12
     AC1 = AC1 >> 13; // Q27 -> Q14
